feat: Implemen Checkout Controller, ShippingController and AddressController

+++ src/main/java/com/ecommerce/webecommerce/model/ShippingOrderRequest.java
@@ -8,6 +8,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Builder
 @Data
 @NoArgsConstructor
@@ -18,7 +20,7 @@ public class ShippingOrderRequest {
     private Long orderId;
     private Address fromAddress;
     private Address toAddress;
-    private int totalWeightInGrams;
+    private BigDecimal totalWeightInGrams;
 
     @Data
     @Builder
